fix(voice): drop the redundant int() cast ruff 0.16 rejects
round() with no ndigits already returns an int, so the wrapping int() is a
no-op. Caught by CI and not locally: requirements-dev.txt pins ruff 0.16.4,
while .pre-commit-config.yaml pins v0.9.6, which predates RUF046.

+++ voice_log.py
@@ -108,4 +108,6 @@ def emit_voice_log(db, background_tasks, uid: str, status: str, **fields) -> Non
 
 def ms(started_at: float, now: float) -> int:
     """Elapsed milliseconds, rounded — the stored twin of the ⏱️ prints."""
-    return int(round((now - started_at) * 1000))
+    # round() with no ndigits already returns an int; wrapping it in int() is
+    # the redundant cast ruff flags as RUF046.
+    return round((now - started_at) * 1000)
